+++ includes/class.Plugin.php
@@ -53,6 +53,22 @@ class Plugin {
 
 		$settings = get_plugin_settings();
 
+		// maybe add an indicator that emails are disabled
+		if ($this->wpmailReplaced) {
+			switch (apply_filters('disable_emails_indicator', $settings['indicator'])) {
+
+				case INDICATOR_TOOLBAR:
+					add_action('admin_bar_menu', [$this, 'showIndicatorToolbar'], 500);
+					add_action('admin_print_styles', [$this, 'styleIndicatorToolbar']);
+					break;
+
+				case INDICATOR_NOTICE:
+					add_action('admin_notices', [$this, 'showIndicatorNotice']);
+					break;
+
+			}
+		}
+
 		// maybe stop BuddyPress emails too
 		if (!empty($settings['buddypress'])) {
 			add_filter('bp_email_use_wp_mail', '__return_true');
@@ -120,6 +136,12 @@ class Plugin {
 			$has_mu_plugin = mu_plugin_manage($_GET['action']);
 		}
 
+		$indicators = [
+			INDICATOR_TOOLBAR		=> _x('Toolbar indicator', 'admin indicator setting', 'disable-emails'),
+			INDICATOR_NOTICE		=> _x('notice on all admin pages', 'admin indicator setting', 'disable-emails'),
+			INDICATOR_NONE			=> _x('no indicator', 'admin indicator setting', 'disable-emails'),
+		];
+
 		require DISABLE_EMAILS_PLUGIN_ROOT . 'views/settings-form.php';
 	}
 
@@ -131,7 +153,12 @@ class Plugin {
 	public function settingsValidate($input) {
 		$output = [];
 
-		$hooknames = [
+		$output['indicator'] = isset($input['indicator']) ? $input['indicator'] : INDICATOR_TOOLBAR;
+		if (!in_array($output['indicator'], [INDICATOR_NONE, INDICATOR_TOOLBAR, INDICATOR_NOTICE])) {
+			add_settings_error(OPT_SETTINGS, 'indicator', _x('Indicator is invalid', 'settings error', 'disable-emails'));
+		}
+
+		$checkboxes = [
 			'wp_mail',
 			'wp_mail_from',
 			'wp_mail_from_name',
@@ -141,7 +168,7 @@ class Plugin {
 			'buddypress',
 			'events_manager',
 		];
-		foreach ($hooknames as $name) {
+		foreach ($checkboxes as $name) {
 			$output[$name] = empty($input[$name]) ? 0 : 1;
 		}
 
@@ -188,23 +215,50 @@ class Plugin {
 	}
 
 	/**
+	* admin notice for indicator of disabled emails status
+	*/
+	public function showIndicatorNotice() {
+		if (current_user_can('activate_plugins') && current_user_can('manage_options')) {
+			include DISABLE_EMAILS_PLUGIN_ROOT . 'views/indicator-notice.php';
+		}
+	}
+
+	/**
+	* Toolbar indicator of disabled emails status
+	* @param WP_Admin_Bar $admin_bar
+	*/
+	public function showIndicatorToolbar($admin_bar) {
+		if (current_user_can('activate_plugins') && current_user_can('manage_options')) {
+			$admin_bar->add_node([
+				'id'		=> 'disable-emails-indicator',
+				'title'		=> sprintf('<span class="ab-icon"></span><span class="screen-reader-text">%s</span>', __('Disable Emails', 'disable-emails')),
+				'href'		=> admin_url('options-general.php?page=disable-emails'),
+				'meta'		=> [
+					'title' => get_status_message(),
+				],
+			]);
+		}
+	}
+
+	public function styleIndicatorToolbar() {
+		?>
+		<style>
+		#wpadminbar #wp-admin-bar-disable-emails-indicator .ab-icon::before {
+			content: "\f465";<?php /* dashicons-email */ ?>
+			top: 3px;
+			background: linear-gradient(to left top, transparent 45%, #dc3232 45%, #dc3232 55%, transparent 60%);
+		}
+		</style>
+		<?php
+	}
+
+	/**
 	* show on admin dashboard that emails have been disabled
 	* @param array $glances
 	*/
 	public function dashboardStatus($glances) {
 		if ($this->wpmailReplaced) {
-			if (defined('DISABLE_EMAILS_MU_PLUGIN') && is_multisite()) {
-				/* translators: shown when emails are disabled for all sites in all networks in a multisite, with the must-use plugin */
-				$dash_msg = __('Emails are disabled for all sites.', 'disable-emails');
-			}
-			elseif (is_plugin_active_for_network(DISABLE_EMAILS_PLUGIN_NAME)) {
-				/* translators: shown when emails are disabled for all sites in a multisite network, by network-activating the plugin */
-				$dash_msg = __('Emails are disabled on this network.', 'disable-emails');
-			}
-			else {
-				/* translators: shown when emails are disabled for the current site */
-				$dash_msg = __('Emails are disabled.', 'disable-emails');
-			}
+			$dash_msg = get_status_message();
 			$glances[] = sprintf('<li style="float:none"><i class="dashicons dashicons-email" aria-hidden="true"></i> %s</li>', esc_html($dash_msg));
 		}