+++ views/settings-form.php
@@ -18,15 +18,27 @@ if (!defined('ABSPATH')) {
 		<table class="form-table">
 
 			<tr valign="top">
+				<th scope="row"><label for="disable_emails_indicator"><?php esc_html_e('Indicator', 'disable-emails'); ?></label></th>
+				<td>
+					<select name="disable_emails[indicator]" id="disable_emails_indicator">
+					<?php foreach($indicators as $value => $label): ?>
+						<option value="<?= esc_attr($value); ?>"<?php selected($value, $settings['indicator']); ?>><?= esc_html($label); ?></option>
+					<?php endforeach; ?>
+					</select>
+					<p><em><?php esc_html_e('Select how you would like the admin to indicate that emails are disabled.', 'disable-emails'); ?></em></p>
+				</td>
+			</tr>
+
+			<tr valign="top">
 				<th scope="row"><?php esc_html_e('Call WordPress hooks', 'disable-emails'); ?></th>
 				<td>
-					<em><?php esc_html_e('call WordPress hooks so that listeners can act, e.g. log emails', 'disable-emails'); ?></em>
-					<br /><label><input type="checkbox" name="disable_emails[wp_mail]" value="1" <?php checked($settings['wp_mail']); ?> /> wp_mail</label>
+					<label><input type="checkbox" name="disable_emails[wp_mail]" value="1" <?php checked($settings['wp_mail']); ?> /> wp_mail</label>
 					<br /><label><input type="checkbox" name="disable_emails[wp_mail_from]" value="1" <?php checked($settings['wp_mail_from']); ?> /> wp_mail_from</label>
 					<br /><label><input type="checkbox" name="disable_emails[wp_mail_from_name]" value="1" <?php checked($settings['wp_mail_from_name']); ?> /> wp_mail_from_name</label>
 					<br /><label><input type="checkbox" name="disable_emails[wp_mail_content_type]" value="1" <?php checked($settings['wp_mail_content_type']); ?> /> wp_mail_content_type</label>
 					<br /><label><input type="checkbox" name="disable_emails[wp_mail_charset]" value="1" <?php checked($settings['wp_mail_charset']); ?> /> wp_mail_charset</label>
 					<br /><label><input type="checkbox" name="disable_emails[phpmailer_init]" value="1" <?php checked($settings['phpmailer_init']); ?> /> phpmailer_init</label>
+					<p><em><?php esc_html_e('call WordPress hooks so that listeners can act, e.g. log emails', 'disable-emails'); ?></em></p>
 				</td>
 			</tr>
 
@@ -56,7 +68,7 @@ if (!defined('ABSPATH')) {
 
 			<p><?= esc_html__('When enabled as a must-use plugin, also known as mu-plugin, Disable Emails is always activated. This is recommended for development websites, in which plugins might be deactivated when refreshing the database from a source website. It can also help when another plugin has already declared wp_mail(), preventing Disable Emails from functioning correctly.', 'disable-emails'); ?>
 
-			<p class="disable-emails-mu-buttons">
+			<div class="disable-emails-mu-buttons">
 				<?php if ($has_mu_plugin): ?>
 
 					<p><?= esc_html__('The must-use plugin is currently enabled.', 'disable-emails'); ?></p>
@@ -78,7 +90,7 @@ if (!defined('ABSPATH')) {
 					<button type="button" class="button button-secondary" id="disable-emails-mu-enable"><?= esc_html__('Activate must-use plugin', 'disable-emails'); ?></button>
 
 				<?php endif; ?>
-			</p>
+			</div>
 
 			<noscript>
 				<p>